requirements.txt
Django>=5.0
psycopg2-binary>=2.9,<3.0
django-environ>=0.7,<1.0
djangorestframework>=3.14